services/prosis-connector/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
oracledb==2.4.1
python-dotenv==1.0.1
pydantic==2.9.2
